fix(prerender): check link's pathname only for extensions
If a link contains parts following the pathname, e.g. a query or hash, the file extension would not be found. Links that appear to have no file extension would then NOT be excluded from crawling. Those included links led the crawler to request, e.g. static, files on which `vue-bundle-renderer` would produce a memory overflow.