Update 5.database.md
The documentation is very centered around local sql db's, I've made a PR on the db0 side of documentation too to include how to connect with Turso. I think it makes a lot of sense to include this here on the Nitro doc section too as a minimum.

Trying to keep duplicate information to a minimum, so I kept the connector- and integration logic at the db0 docs.